ajustes.

+++ src/Company.php
@@ -71,7 +71,7 @@ class Company
         } catch (\Exception $e) {
             throw new \Exception("Erro ao Conectar com SAP: " . $e->getMessage());
         }
-        $this->_com->DbServerType = config("sap.db.type");
+        $this->_com->DbServerType = (int)config("sap.db.type");
         $this->_com->Server = config("sap.server");
         $this->_com->LicenseServer = config("sap.license_server");
         $this->_com->UserName = config("sap.username");
@@ -79,7 +79,7 @@ class Company
         $this->_com->CompanyDB = config("sap.db.name");
         $this->_com->DbUserName = config("sap.db.username");
         $this->_com->DbPassword = config("sap.db.password");
-        $this->_com->language = config("sap.db.language");
+        $this->_com->language = (int)config("sap.db.language");
         $retVal = $this->_com->Connect();
 
         if ($retVal != "0") {
